refactor: split up utils
