+++ packages/common/src/index.ts
@@ -14,9 +14,6 @@ export { BaseError } from "./errors/BaseError.js";
 export { ChainNotFoundError } from "./errors/ChainNotFoundError.js";
 export { AccountNotFoundError } from "./errors/AccountNotFoundError.js";
 export { FetchError } from "./errors/FetchError.js";
-export { InvalidDeferredActionNonceError } from "./errors/InvalidDeferredActionNonceError.js";
-export { InvalidEntityIdError } from "./errors/InvalidEntityIdError.js";
-export { InvalidNonceKeyError } from "./errors/InvalidNonceKeyError.js";
 export { ServerError } from "./errors/ServerError.js";
 
 // actions
