+++ packages/smart-accounts/src/ma-v2/modules/allowlist-module/abis/allowlistModuleAbi.ts
