+++ packages/smart-accounts/src/ma-v2/modules/single-signer-validation/abis/singleSignerValidationModuleAbi.ts
